// PRO-EP/pro-ep-ui/src/app/interfaces/interfaces.ts
export interface IUserMin {
  userid: number;
  dempoid: string;
  displayName: string | null;
  fname: string;
  lname: string;
  preferredfname: string;
  preferredlname: string;
  language: string;
  role: number;
  buddy?: boolean;
  employmenttype?: number;
  trainedon: string;
  trainedOnArray: string[];
  active?: boolean;
  canEdit?: boolean;
  selected?: boolean;
  checked?: boolean;
  changed?: boolean;
  schedulinglevel?: number;
}

export interface ITimeCard {
  timeCardId: number;
  dempoid: string;
  //workDate?: Date;
  timeIn?: Date | null;
  timeOut?: Date | null;
  entryBy: string;
  entryDt?: Date;
  modBy: string;
  modDt?: Date;
  datetimein?: Date | null;
  datetimeout?: Date | null;
  externalIp: string | null;
  //MachineName: string;
}

export interface ICurrentUser {
  user: IUser;
  timecards: ITimeCard[];
}

export interface IProject {
  projectID: number | null;
  projectName: string | null;
  projectLeadId?: number | null;
  projectLead: string | null;
  projectCoordinatorId?: number | null;
  projectCoordinator: string | null;
  active?: number | null;
  projectStatus: string | null;
  projectDisplayId: string | null;
  interview?: number | null;
  projectColor: string | null;
  projectColorIview: string | null;
  projectedStartDate?: Date | null;
  projectedEndDate?: Date | null;
  projectEpmcode: string | null;
  projectType?: number | null;
  sponsor: string | null;
  principalInvestigator: string | null;
  totalBudget?: number | null;
  comments: string | null;
  scheduleDisplay?: number | null;
  ssdatabaseName: string | null;
  forecast?: number | null;
  tollFreeNumber: string | null;
  studyEmailAddress: string | null;
  pdriveLocation: string | null;
  tdriveLocation: string | null;
  irbnumber: string | null;
  fundCode: string | null;
  sisterCode: string | null;
  passThruCode: string | null;
  rescue?: number | null;
  orphan?: number | null;
  edcsystem: string | null;
  therapeuticArea: string | null;
  projectAbbr: string | null;
  projectManager: string | null;
  faxNumber: string | null;
  faxLocation: string | null;
  maxEnrollment?: number | null;
  firstEnrollDate?: Date | null;
  visitDate: string | null;
  totalProjected?: number | null;
  actualFollowUp?: number | null;
  numofIntervals?: number;
  entryFormName: string | null;
  entryDt: Date | null;
  entryBy: string | null;
  modDt: Date | null;
  modBy: string | null;
  trainedOnUsersMin?: IUserMin[];
  trainedOnUsers?: string[];
  notTrainedOnUsers?: string[];
}

export interface IProjectMin {
  projectID: number;
  projectName: string;
  projectAbbr: string;
  projectDisplayId: string;
  displayedIn?: string;
  active?: boolean | null;
  projectStatus?: boolean | null;
  projectColor: string;
  projectType?: string | null;
  selected?: boolean;
  trainedOnUsersMin?: IUserMin[];
  trainedOnUsers?: string[];
  notTrainedOnUsers?: string[];
}

export interface IDropDownValue {
  dropDownValueId?: number;
  sortOrder?: number;
  codeValues?: number;
  formFieldId?: number;
  dropDownItem?: string;
  colorCode?: string | null;
  abbr?: string | null;
  hidden?: boolean;
  confirmDelete?: boolean;
  tempKey?: number;
  changed?: boolean;
  invalid?: boolean;
  entryDt?: Date | null;
  entryBy?: string | null;
  modDt?: Date | null;
  modBy?: string | null;
}

export interface IFormField {
  formFieldId: number;
  fieldLabel: string;
  defaultLabel: string;
  columnName: string;
  tableName: string;
  formOrder: number;
  tab: string | null;
  fieldType: string;
  fieldTypeSpecific?: string;
  configurable: string;
  wasRequired?: boolean;
  required?: boolean;
  displayInTable?: boolean;
  hidden?: boolean;
  projectId?: number;
  requireConfirmation?: boolean;
  entryDt?: Date;
  entryBy: string;
  modDt?: Date;
  modBy: string;
}

export interface IAdminOptionsVariable {
  adminOptionsId: number;
  optionValue: string;
  optionValueMulti: Array<string>;
  fieldType: string;
  sortOrder: number;
  fieldLabel: string;
  dropDownValues?: Array<IDropDownValue>;
  formFieldId: number;
}

export interface IFormFieldVariable {
  formField: IFormField;
  dropDownValues?: Array<IDropDownValue>;
  expanded?: boolean;
  changed?: boolean;
  invalid?: boolean;
  filterValue?: any;
  sortActive?: boolean;
}

export interface IFormFieldInstance {
  value: any;
  formFieldVariable: IFormFieldVariable;
  invalid: boolean;
  missingRequired?: boolean;
  validationError?: boolean;
  validationMessage: string;
}
// ...
export interface IAuthenticatedUser {
  displayName: string;
  duDukeID: string;
  eppn: string;
  netID: string;
  isMemberOf: string;
  interviewer: boolean;
  resourceGroup: boolean;
  admin: boolean;
  userRoles?: string[] | undefined;
  sessionMinsLeft: number;
  timecards: ITimeCard[];
}
// ...
export interface IForecastHours {
  forecastHoursId: number;
  projectId?: number;
  month1?: Date;
  forecastHours1?: number;
  comment1: string;
  weekchk1?: boolean;
  weekendChk1?: boolean;
  freq1?: number;
  month2?: Date;
  forecastHours2?: number;
  comment2: string;
  weekchk2?: boolean;
  weekendChk2?: boolean;
  freq2?: number;
  month3?: Date;
  forecastHours3?: number;
  comment3: string;
  weekchk3?: boolean;
  weekendChk3?: boolean;
  freq3?: number;
  month4?: Date;
  forecastHours4?: number;
  comment4: string;
  weekchk4?: boolean;
  weekendChk4?: boolean;
  freq4?: number;
  month5?: Date;
  forecastHours5?: number;
  comment5: string;
  weekchk5?: boolean;
  weekendChk5?: boolean;
  freq5?: number;
  month6?: Date;
  forecastHours6?: number;
  comment6: string;
  weekchk6?: boolean;
  weekendChk6?: boolean;
  freq6?: number;
  month7?: Date;
  forecastHours7?: number;
  comment7: string;
  weekchk7?: boolean;
  weekendChk7?: boolean;
  freq7?: number;
  month8?: Date;
  forecastHours8?: number;
  comment8: string;
  weekchk8?: boolean;
  weekendChk8?: boolean;
  freq8?: number;
  month9?: Date;
  forecastHours9?: number;
  comment9: string;
  weekchk9?: boolean;
  weekendChk9?: boolean;
  freq9?: number;
  month10?: Date;
  forecastHours10?: number;
  comment10: string;
  weekchk10?: boolean;
  weekendChk10?: boolean;
  freq10?: number;
  month11?: Date;
  forecastHours11?: number;
  comment11: string;
  weekchk11?: boolean;
  weekendChk11?: boolean;
  freq11?: number;
  month12?: Date;
  forecastHours12?: number;
  comment12: string;
  weekchk12?: boolean;
  weekendChk12?: boolean;
  freq12?: number;
  month13?: Date;
  forecastHours13?: number;
  comment13: string;
  weekchk13?: boolean;
  weekendChk13?: boolean;
  freq13?: number;
  month14?: Date;
  forecastHours14?: number;
  comment14: string;
  weekchk14?: boolean;
  weekendChk14?: boolean;
  freq14?: number;
  entryBy: string;
  entryDt?: Date;
  modBy: string;
  modDt?: Date;
}